requirements.txt
asgiref==3.8.1
Django==4.2.20
sqlparse==0.5.3
typing_extensions==4.13.2
# Core dependencies
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers
# Backend global dependencies
Pillow>=9.0.0
django-unfold